+++ tools/filamesh/README.md
@@ -29,7 +29,7 @@ Note: the UV1 attribute cannot be used in interleaved mode
     uint32  : number of parts (sub-meshes or draw calls)
     float3  : center of the total bounding box (AABB)
     float3  : half extent of the total bounding box (AABB)
-    uint32  : 0 if the vertices are not interleaved, 1 otherwise
+    uint32  : flags (interleaved, UV type, compression type)
     uint32  : offset of the position attribute
     uint32  : stride of the position attribute
     uint32  : offset of the tangents attribute
@@ -97,7 +97,7 @@ struct Header {
     uint32_t version;
     uint32_t parts;
     Box      aabb;
-    uint32_t interleaved;
+    uint32_t flags;
     uint32_t offsetPosition;
     uint32_t stridePosition;
     uint32_t offsetTangents;
@@ -119,7 +119,7 @@ struct Vertex {
     half4  position;
     short4 tangents;
     ubyte4 color;
-    half2  uv0;
+    short2 uv0; // either half-float or snorm int16
 };
 
 struct Part {
